object space zoom

+++ src/transform.c
@@ -6,7 +6,7 @@
 /*   By: tspoof <tspoof@student.hive.fi>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/03/23 14:26:10 by tspoof            #+#    #+#             */
-/*   Updated: 2023/03/26 17:12:30 by tspoof           ###   ########.fr       */
+/*   Updated: 2023/03/28 13:42:03 by tspoof           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -18,8 +18,8 @@ static void	do_transform(void *dst, void *params, t_map *map, t_xy xy)
 
 	(void)dst;
 	transforms = (t_xy *)params;
-	map->grid[xy.y][xy.x].x = map->grid[xy.y][xy.x].x + transforms->x;
-	map->grid[xy.y][xy.x].y = map->grid[xy.y][xy.x].y + transforms->y;
+	map->grid[xy.y][xy.x].x = map->grid[xy.y][xy.x].x + (float)transforms->x;
+	map->grid[xy.y][xy.x].y = map->grid[xy.y][xy.x].y + (float)transforms->y;
 }
 
 void	ft_transform(t_map *map, int transform_x, int transform_y)
